requirements.txt
spotipy==2.21.0
python-dotenv==1.0.0